refactor CGRA and multi CGRA model, move them out from main UI file
